Finish terminal updates
