+++ src/terminal/terminal.ts
@@ -1,0 +1,294 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import {
+  IIterator
+} from 'phosphor/lib/algorithm/iteration';
+
+import {
+  JSONPrimitive, JSONObject
+} from 'phosphor/lib/algorithm/json';
+
+import {
+  IDisposable
+} from 'phosphor/lib/core/disposable';
+
+import {
+  ISignal
+} from 'phosphor/lib/core/signaling';
+
+import {
+  IAjaxSettings
+} from '../utils';
+
+import * as utils
+  from '../utils';
+
+import {
+  DefaultTerminalSession
+} from './default';
+
+
+/**
+ * The namespace for TerminalSession.ISession statics.
+ */
+export
+namespace TerminalSession {
+  /**
+   * An interface for a terminal session.
+   */
+  export
+  interface ISession extends IDisposable {
+    /**
+     * A signal emitted when the session is shut down.
+     */
+    terminated: ISignal<ISession, void>;
+
+    /**
+     * A signal emitted when a message is received from the server.
+     */
+    messageReceived: ISignal<ISession, IMessage>;
+
+    /**
+     * Get the name of the terminal session.
+     */
+    readonly name: string;
+
+    /**
+     * The model associated with the session.
+     */
+    readonly model: IModel;
+
+    /**
+     * The base url of the session.
+     */
+    readonly baseUrl: string;
+
+    /**
+     * The Ajax settings used for server requests.
+     */
+    ajaxSettings: utils.IAjaxSettings;
+
+    /**
+     * Send a message to the terminal session.
+     */
+    send(message: IMessage): void;
+
+    /**
+     * Shut down the terminal session.
+     */
+    shutdown(): Promise<void>;
+  }
+
+  /**
+   * Start a new terminal session.
+   *
+   * @options - The session options to use.
+   *
+   * @returns A promise that resolves with the session instance.
+   */
+  export
+  function startNew(options?: IOptions): Promise<ISession> {
+    return DefaultTerminalSession.startNew(options);
+  }
+
+  /*
+   * Connect to a running session.
+   *
+   * @param name - The name of the target session.
+   *
+   * @param options - The session options to use.
+   *
+   * @returns A promise that resolves with the new session instance.
+   *
+   * #### Notes
+   * If the session was already started via `startNew`, the existing
+   * session object is used as the fulfillment value.
+   *
+   * Otherwise, if `options` are given, we attempt to connect to the existing
+   * session.
+   * The promise is fulfilled when the session is ready on the server,
+   * otherwise the promise is rejected.
+   *
+   * If the session was not already started and no `options` are given,
+   * the promise is rejected.
+   */
+  export
+  function connectTo(name: string, options?: IOptions): Promise<ISession> {
+    return DefaultTerminalSession.connectTo(name, options);
+  }
+
+  /**
+   * List the running terminal sessions.
+   *
+   * @param options - The session options to use.
+   *
+   * @returns A promise that resolves with the list of running session models.
+   */
+  export
+  function listRunning(options: IOptions): Promise<IModel[]> {
+    return DefaultTerminalSession.listRunning(options);
+  }
+
+  /**
+   * Shut down a terminal session by name.
+   *
+   * @param name - The name of the target session.
+   *
+   * @param options - The session options to use.
+   *
+   * @returns A promise that resolves when the session is shut down.
+   */
+  export
+  function shutdown(name: string, options?: IOptions): Promise<void> {
+    return DefaultTerminalSession.shutdown(name, options);
+  }
+
+  /**
+   * The options for intializing a terminal session object.
+   */
+  export
+  interface IOptions extends JSONObject {
+    /**
+     * The name of the terminal session.
+     */
+    name?: string;
+
+    /**
+     * The base url.
+     */
+    baseUrl?: string;
+
+    /**
+     * The base websocket url.
+     */
+    wsUrl?: string;
+
+    /**
+     * The Ajax settings used for server requests.
+     */
+    ajaxSettings?: utils.IAjaxSettings;
+  }
+
+  /**
+   * The server model for a terminal session.
+   */
+  export
+  interface IModel extends JSONObject {
+    /**
+     * The name of the terminal session.
+     */
+    readonly name: string;
+  }
+
+  /**
+   * A message from the terminal session.
+   */
+  export
+  interface IMessage {
+    /**
+     * The type of the message.
+     */
+    readonly type: MessageType;
+
+    /**
+     * The content of the message.
+     */
+    readonly content?: JSONPrimitive[];
+  }
+
+  /**
+   * Valid message types for the terminal.
+   */
+  export
+  type MessageType = 'stdout' | 'disconnect' | 'set_size' | 'stdin';
+
+  /**
+   * The interface for a terminal manager.
+   *
+   * #### Notes
+   * The manager is responsible for maintaining the state of running
+   * terminal sessions.
+   */
+  export
+  interface IManager extends IDisposable {
+    /**
+     * A signal emitted when the running terminals change.
+     */
+    runningChanged: ISignal<IManager, IModel[]>;
+
+    /**
+     * The base url of the manager.
+     */
+    readonly baseUrl: string;
+
+    /**
+     * The base ws url of the manager.
+     */
+    readonly wsUrl: string;
+
+    /**
+     * The default ajax settings for the manager.
+     */
+    ajaxSettings?: IAjaxSettings;
+
+    /**
+     * Create an iterator over the known running terminals.
+     *
+     * @returns A new iterator over the running terminals.
+     */
+    running(): IIterator<IModel>;
+
+    /**
+     * Create a new terminal session.
+     *
+     * @param ajaxSettings - The ajaxSettings to use, overrides manager
+     *   settings.
+     *
+     * @returns A promise that resolves with the terminal instance.
+     *
+     * #### Notes
+     * The baseUrl and wsUrl of the options will be forced
+     * to the ones used by the manager. The ajaxSettings of the manager
+     * will be used unless overridden.
+     */
+    startNew(options?: TerminalSession.IOptions): Promise<TerminalSession.ISession>;
+
+    /*
+     * Connect to a running session.
+     *
+     * @param name - The name of the target session.
+     *
+     * @param ajaxSettings - The ajaxSettings to use, overrides manager
+     *   settings.
+     *
+     * @returns A promise that resolves with the new session instance.
+     *
+     * #### Notes
+     * The baseUrl and wsUrl of the options will be forced
+     * to the ones used by the manager. The ajaxSettings of the manager
+     * will be used unless overridden.
+     */
+    connectTo(name: string, options?: TerminalSession.IOptions): Promise<TerminalSession.ISession>;
+
+    /**
+     * Shut down a terminal session by name.
+     *
+     * @param name - The name of the terminal session.
+     *
+     * @returns A promise that resolves when the session is shut down.
+     */
+    shutdown(name: string): Promise<void>;
+
+    /**
+     * Force a refresh of the running terminals.
+     *
+     * @returns A promise that resolves when the refresh is complete.
+     *
+     * #### Notes
+     * This is not typically meant to be called by the user, since the
+     * manager maintains its own internal state.
+     */
+    refreshRunning(): Promise<void>;
+  }
+}
